Add Time module
