+++ hvega/src/Graphics/Vega/VegaLite.hs
@@ -608,6 +608,7 @@ import qualified Graphics.Vega.VegaLite.Core as VL
 import qualified Graphics.Vega.VegaLite.Output as VL
 import qualified Graphics.Vega.VegaLite.Selection as VL
 import qualified Graphics.Vega.VegaLite.Specification as VL
+import qualified Graphics.Vega.VegaLite.Time as VL
 
 
 -- Documentation
